+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### Version 9.2
+* Adds `FallbackFactory`, allowing access to the cause of a Hystrix fallback
 * Supports context path when using Ribbon `LoadBalancingTarget`
 * Adds builder methods for the Response object
 * Deprecates Response factory methods